requirements.txt
flask>=3.0.0
pymupdf>=1.24.0
openpyxl>=3.1.0
Pillow>=10.0.0
camelot-py>=0.11.0
pdfplumber>=0.11.0
tabula-py>=2.9.0
pandas>=2.0.0
numpy>=1.26.0
pytesseract>=0.3.10
requests>=2.28.0
convertapi>=2.0.0
python-dotenv>=1.0.0
gunicorn>=22.0.0
